requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
Werkzeug==2.3.7
